Add comments explaining forward selection
